Migrate all model types into the API files.
This will make it easier to iterate on APIs by keeping the models close to the APIs that use them.
