+++ Sources/CommandLineTool/Command+ShowStatus.swift
@@ -76,7 +76,7 @@ extension Porsche {
       }
     }
 
-    private func formatted(distance: Status.Distance, scalar: Double = 1) -> String {
+    private func formatted(distance: Distance, scalar: Double = 1) -> String {
       let formatter = MeasurementFormatter()
       formatter.unitStyle = .long
       formatter.unitOptions = .providedUnit
@@ -84,12 +84,10 @@ extension Porsche {
       let value = distance.value * scalar
       let unit: UnitLength
       switch distance.unit {
-      case "KILOMETERS":
+      case .kilometers:
         unit = .kilometers
-      case "MILES":
+      case .miles:
         unit = .miles
-      default:
-        return "\(value)"
       }
       return formatter.string(from: Measurement(value: value, unit: unit))
     }
