+++ Sources/PorscheConnect/APIs/PorscheConnect+Auth.swift
@@ -121,3 +121,16 @@ extension PorscheConnect {
     ]
   }
 }
+
+// MARK: - Response types
+
+/// A response from one of the Porsche Connect authorization endpoints.
+///
+/// This type is not meant to be stored to disk as it includes a relative time value that is only meaningful when
+/// first decoded from the server. If you need to store an AuthResponse longer-term, use OAuthToken instead.
+struct AuthResponse: Decodable {
+  let accessToken: String
+  let idToken: String
+  let tokenType: String
+  let expiresIn: Double
+}
