+++ Sources/PorscheConnect/APIs/PorscheConnect+Position.swift
@@ -1,3 +1,4 @@
+import CoreLocation
 import Foundation
 
 extension PorscheConnect {
@@ -10,5 +11,26 @@ extension PorscheConnect {
       Position.self, url: networkRoutes.vehiclePositionURL(vin: vin), headers: headers,
       jsonKeyDecodingStrategy: .useDefaultKeys)
     return (position: result.data, response: result.response)
+  }
+}
+
+// MARK: - Response types
+
+public struct Position: Codable {
+
+  // MARK: Properties
+
+  public let carCoordinate: CarCoordinate
+  public let heading: CLLocationDirection
+
+  // MARK: -
+
+  public struct CarCoordinate: Codable {
+
+    // MARK: Properties
+
+    public let geoCoordinateSystem: String
+    public let latitude: CLLocationDegrees
+    public let longitude: CLLocationDegrees
   }
 }
