--- Sources/PorscheConnect/Models/Summary.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-public struct Summary: Codable {
-
-  // MARK: Properties
-
-  public let modelDescription: String
-  public let nickName: String?
-}
