+++ Tests/PorscheConnectTests/Public/PorscheConnect+TripsTests.swift
@@ -49,21 +49,21 @@ final class PorscheConnectTripTests: BaseMockNetworkTestCase {
     
     let averageSpeed = trip.averageSpeed
     XCTAssertEqual(11, averageSpeed.value)
-    XCTAssertEqual(SpeedUnit.kmh, averageSpeed.unit)
+    XCTAssertEqual(Speed.Unit.kmh, averageSpeed.unit)
     XCTAssertEqual(11, averageSpeed.valueInKmh)
     XCTAssertEqual("TC.UNIT.KMH", averageSpeed.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KMH", averageSpeed.unitTranslationKey)
     
     let averageFuelConsumption = trip.averageFuelConsumption
     XCTAssertEqual(0, averageFuelConsumption.value)
-    XCTAssertEqual(Trip.AverageFuelConsumption.FuelConsumptionUnit.litersPer100Km, averageFuelConsumption.unit)
+    XCTAssertEqual(FuelConsumption.Unit.litersPer100Km, averageFuelConsumption.unit)
     XCTAssertEqual(0, averageFuelConsumption.valueInLitersPer100Km)
     XCTAssertEqual("TC.UNIT.LITERS_PER_100_KM", averageFuelConsumption.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_LITERS_PER_100_KM", averageFuelConsumption.unitTranslationKey)
     
     let averageElectricEngineConsumption = trip.averageElectricEngineConsumption
     XCTAssertEqual(39.6, averageElectricEngineConsumption.value)
-    XCTAssertEqual(Trip.AverageElectricEngineConsumption.ElectricConsumptionUnit.kilowattHoursPer100Km, averageElectricEngineConsumption.unit)
+    XCTAssertEqual(ElectricEngineConsumption.Unit.kilowattHoursPer100Km, averageElectricEngineConsumption.unit)
     XCTAssertEqual(39.6, averageElectricEngineConsumption.valueKwhPer100Km)
     XCTAssertEqual("TC.UNIT.KWH_PER_100KM", averageElectricEngineConsumption.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KWH_PER_100KM", averageElectricEngineConsumption.unitTranslationKey)
@@ -72,8 +72,8 @@ final class PorscheConnectTripTests: BaseMockNetworkTestCase {
     XCTAssertEqual(6, tripMileage.value)
     XCTAssertEqual(6, tripMileage.originalValue)
     XCTAssertEqual(6, tripMileage.valueInKilometers)
-    XCTAssertEqual(DistanceUnit.kilometers, tripMileage.unit)
-    XCTAssertEqual(DistanceUnit.kilometers, tripMileage.originalUnit)
+    XCTAssertEqual(Distance.Unit.kilometers, tripMileage.unit)
+    XCTAssertEqual(Distance.Unit.kilometers, tripMileage.originalUnit)
     XCTAssertEqual("TC.UNIT.KILOMETER", tripMileage.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KILOMETER", tripMileage.unitTranslationKey)
     
@@ -81,8 +81,8 @@ final class PorscheConnectTripTests: BaseMockNetworkTestCase {
     XCTAssertEqual(1442, startMileage.value)
     XCTAssertEqual(1442, startMileage.originalValue)
     XCTAssertEqual(1442, startMileage.valueInKilometers)
-    XCTAssertEqual(DistanceUnit.kilometers, startMileage.unit)
-    XCTAssertEqual(DistanceUnit.kilometers, startMileage.originalUnit)
+    XCTAssertEqual(Distance.Unit.kilometers, startMileage.unit)
+    XCTAssertEqual(Distance.Unit.kilometers, startMileage.originalUnit)
     XCTAssertEqual("TC.UNIT.KILOMETER", startMileage.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KILOMETER", startMileage.unitTranslationKey)
     
@@ -90,8 +90,8 @@ final class PorscheConnectTripTests: BaseMockNetworkTestCase {
     XCTAssertEqual(1448, endMileage.value)
     XCTAssertEqual(1448, endMileage.originalValue)
     XCTAssertEqual(1448, endMileage.valueInKilometers)
-    XCTAssertEqual(DistanceUnit.kilometers, endMileage.unit)
-    XCTAssertEqual(DistanceUnit.kilometers, endMileage.originalUnit)
+    XCTAssertEqual(Distance.Unit.kilometers, endMileage.unit)
+    XCTAssertEqual(Distance.Unit.kilometers, endMileage.originalUnit)
     XCTAssertEqual("TC.UNIT.KILOMETER", endMileage.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KILOMETER", endMileage.unitTranslationKey)
     
@@ -99,8 +99,8 @@ final class PorscheConnectTripTests: BaseMockNetworkTestCase {
     XCTAssertEqual(6, zeroEmissionDistance.value)
     XCTAssertEqual(6, zeroEmissionDistance.originalValue)
     XCTAssertEqual(6, zeroEmissionDistance.valueInKilometers)
-    XCTAssertEqual(DistanceUnit.kilometers, zeroEmissionDistance.unit)
-    XCTAssertEqual(DistanceUnit.kilometers, zeroEmissionDistance.originalUnit)
+    XCTAssertEqual(Distance.Unit.kilometers, zeroEmissionDistance.unit)
+    XCTAssertEqual(Distance.Unit.kilometers, zeroEmissionDistance.originalUnit)
     XCTAssertEqual("TC.UNIT.KILOMETER", zeroEmissionDistance.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KILOMETER", zeroEmissionDistance.unitTranslationKey)
 
@@ -145,21 +145,21 @@ final class PorscheConnectTripTests: BaseMockNetworkTestCase {
     
     let averageSpeed = trip.averageSpeed
     XCTAssertEqual(39, averageSpeed.value)
-    XCTAssertEqual(SpeedUnit.kmh, averageSpeed.unit)
+    XCTAssertEqual(Speed.Unit.kmh, averageSpeed.unit)
     XCTAssertEqual(39, averageSpeed.valueInKmh)
     XCTAssertEqual("TC.UNIT.KMH", averageSpeed.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KMH", averageSpeed.unitTranslationKey)
     
     let averageFuelConsumption = trip.averageFuelConsumption
     XCTAssertEqual(0, averageFuelConsumption.value)
-    XCTAssertEqual(Trip.AverageFuelConsumption.FuelConsumptionUnit.litersPer100Km, averageFuelConsumption.unit)
+    XCTAssertEqual(FuelConsumption.Unit.litersPer100Km, averageFuelConsumption.unit)
     XCTAssertEqual(0, averageFuelConsumption.valueInLitersPer100Km)
     XCTAssertEqual("TC.UNIT.LITERS_PER_100_KM", averageFuelConsumption.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_LITERS_PER_100_KM", averageFuelConsumption.unitTranslationKey)
     
     let averageElectricEngineConsumption = trip.averageElectricEngineConsumption
     XCTAssertEqual(29.9, averageElectricEngineConsumption.value)
-    XCTAssertEqual(Trip.AverageElectricEngineConsumption.ElectricConsumptionUnit.kilowattHoursPer100Km, averageElectricEngineConsumption.unit)
+    XCTAssertEqual(ElectricEngineConsumption.Unit.kilowattHoursPer100Km, averageElectricEngineConsumption.unit)
     XCTAssertEqual(29.9, averageElectricEngineConsumption.valueKwhPer100Km)
     XCTAssertEqual("TC.UNIT.KWH_PER_100KM", averageElectricEngineConsumption.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KWH_PER_100KM", averageElectricEngineConsumption.unitTranslationKey)
@@ -168,8 +168,8 @@ final class PorscheConnectTripTests: BaseMockNetworkTestCase {
     XCTAssertEqual(1448, tripMileage.value)
     XCTAssertEqual(1448, tripMileage.originalValue)
     XCTAssertEqual(1448, tripMileage.valueInKilometers)
-    XCTAssertEqual(DistanceUnit.kilometers, tripMileage.unit)
-    XCTAssertEqual(DistanceUnit.kilometers, tripMileage.originalUnit)
+    XCTAssertEqual(Distance.Unit.kilometers, tripMileage.unit)
+    XCTAssertEqual(Distance.Unit.kilometers, tripMileage.originalUnit)
     XCTAssertEqual("TC.UNIT.KILOMETER", tripMileage.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KILOMETER", tripMileage.unitTranslationKey)
     
@@ -177,8 +177,8 @@ final class PorscheConnectTripTests: BaseMockNetworkTestCase {
     XCTAssertEqual(-1, startMileage.value)
     XCTAssertEqual(-1, startMileage.originalValue)
     XCTAssertEqual(-1, startMileage.valueInKilometers)
-    XCTAssertEqual(DistanceUnit.kilometers, startMileage.unit)
-    XCTAssertEqual(DistanceUnit.kilometers, startMileage.originalUnit)
+    XCTAssertEqual(Distance.Unit.kilometers, startMileage.unit)
+    XCTAssertEqual(Distance.Unit.kilometers, startMileage.originalUnit)
     XCTAssertEqual("TC.UNIT.KILOMETER", startMileage.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KILOMETER", startMileage.unitTranslationKey)
     
@@ -186,8 +186,8 @@ final class PorscheConnectTripTests: BaseMockNetworkTestCase {
     XCTAssertEqual(1448, endMileage.value)
     XCTAssertEqual(1448, endMileage.originalValue)
     XCTAssertEqual(1448, endMileage.valueInKilometers)
-    XCTAssertEqual(DistanceUnit.kilometers, endMileage.unit)
-    XCTAssertEqual(DistanceUnit.kilometers, endMileage.originalUnit)
+    XCTAssertEqual(Distance.Unit.kilometers, endMileage.unit)
+    XCTAssertEqual(Distance.Unit.kilometers, endMileage.originalUnit)
     XCTAssertEqual("TC.UNIT.KILOMETER", endMileage.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KILOMETER", endMileage.unitTranslationKey)
     
@@ -195,8 +195,8 @@ final class PorscheConnectTripTests: BaseMockNetworkTestCase {
     XCTAssertEqual(1448, zeroEmissionDistance.value)
     XCTAssertEqual(1448, zeroEmissionDistance.originalValue)
     XCTAssertEqual(1448, zeroEmissionDistance.valueInKilometers)
-    XCTAssertEqual(DistanceUnit.kilometers, zeroEmissionDistance.unit)
-    XCTAssertEqual(DistanceUnit.kilometers, zeroEmissionDistance.originalUnit)
+    XCTAssertEqual(Distance.Unit.kilometers, zeroEmissionDistance.unit)
+    XCTAssertEqual(Distance.Unit.kilometers, zeroEmissionDistance.originalUnit)
     XCTAssertEqual("TC.UNIT.KILOMETER", zeroEmissionDistance.unitTranslationKeyV2)
     XCTAssertEqual("GRAY_SLICE_UNIT_KILOMETER", zeroEmissionDistance.unitTranslationKey)
